+++ 3ISIP223_PogosyanWPF/DopOptinon.cs
@@ -9,9 +9,9 @@ namespace _3ISIP223_PogosyanWPF
     internal class DopOptinon
     {
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
-        public DopOptinon(string name, decimal price)
+        public DopOptinon(string name, double price)
         {
             Name = name;    
             Price = price;
